+++ library/PSFramework/Parameter/PathLiteralFileSingleParameter.cs
@@ -9,45 +9,45 @@ using System.Threading.Tasks;
 namespace PSFramework.Parameter
 {
     /// <summary>
-    /// Parameterclass that accepts a single directory
+    /// Parameter Class that maps to a single file without processing wildcards
     /// </summary>
-    public class PathDirectorySingleParameter : PathFileSystemSingleParameterBase
+    public class PathLiteralFileSingleParameter : PathFileSystemSingleParameterBase
     {
         #region Constructors
         /// <summary>
-        /// Processes a single string as a single directory.
+        /// Processes a single string as a single file.
         /// </summary>
         /// <param name="Path">The path to process</param>
-        public PathDirectorySingleParameter(string Path)
+        public PathLiteralFileSingleParameter(string Path)
         {
             InputObject = Path;
-            Apply(Path, false, true);
+            ApplyLiteral(Path, true, false);
         }
 
         /// <summary>
-        /// Processes a single DirectoryInfo item as a single directory.
+        /// Processes a single FileInfo item as a single file.
         /// </summary>
         /// <param name="Path">The path to process</param>
-        public PathDirectorySingleParameter(DirectoryInfo Path) : this(Path.FullName) { InputObject = Path; }
+        public PathLiteralFileSingleParameter(FileInfo Path) : this(Path.FullName) { InputObject = Path; }
 
         /// <summary>
-        /// Processes a single Uri as a single directory.
+        /// Processes a single Uri as a single file.
         /// </summary>
         /// <param name="Path">The path to process</param>
-        public PathDirectorySingleParameter(Uri Path) : this(Path.OriginalString) { InputObject = Path; }
+        public PathLiteralFileSingleParameter(Uri Path) : this(Path.OriginalString) { InputObject = Path; }
 
         /// <summary>
-        /// Processes a single object as a single directory.
+        /// Processes a single object as a single file.
         /// </summary>
         /// <param name="Path">The path to process</param>
-        public PathDirectorySingleParameter(object Path)
+        public PathLiteralFileSingleParameter(object Path)
         {
             InputObject = Path;
             string actualpath;
             try { actualpath = LanguagePrimitives.ConvertTo<string>(PathFileSystemParameterBase.GetObject(Path)); }
             catch (Exception e) { throw new ArgumentException($"Failed to process {Path}! Error converting to string: {e.Message}", e); }
 
-            Apply(actualpath, false, true);
+            ApplyLiteral(actualpath, true, false);
         }
         #endregion Constructors
 
@@ -56,35 +56,36 @@ namespace PSFramework.Parameter
         /// Implicitly convert to string.
         /// </summary>
         /// <param name="Path">The path to convert</param>
-        public static implicit operator string(PathDirectorySingleParameter Path)
+        public static implicit operator string(PathLiteralFileSingleParameter Path)
         {
             return Path.Path;
         }
+
         /// <summary>
-        /// Implicitly convert to DirectoryInfo.
+        /// Implicitly convert to FileInfo.
         /// </summary>
         /// <param name="Path">The path to convert</param>
-        public static implicit operator DirectoryInfo(PathDirectorySingleParameter Path)
+        public static implicit operator FileInfo(PathLiteralFileSingleParameter Path)
         {
-            return new DirectoryInfo(Path.Path);
+            return new FileInfo(Path.Path);
         }
+
         /// <summary>
-        /// Implicitly convert DirectoryInfo to PathDirectorySingleParameter.
+        /// Implicitly convert FileInfo to PathLiteralFileSingleParameter.
         /// </summary>
         /// <param name="Info">The path to convert</param>
-
-        public static implicit operator PathDirectorySingleParameter(DirectoryInfo Info)
+        public static implicit operator PathLiteralFileSingleParameter(FileInfo Info)
         {
-            return new PathDirectorySingleParameter(Info);
+            return new PathLiteralFileSingleParameter(Info);
         }
+
         /// <summary>
         /// Implicitly convert to FileSystemInfo.
         /// </summary>
         /// <param name="Path">The path to convert</param>
-
-        public static implicit operator FileSystemInfo(PathDirectorySingleParameter Path)
+        public static implicit operator FileSystemInfo(PathLiteralFileSingleParameter Path)
         {
-            return new DirectoryInfo(Path.Path);
+            return new FileInfo(Path.Path);
         }
         #endregion Operators
     }